Moved App component and related css file into the corresponding folders. Fixed imports.
